refactor: move functions into better subfolders
