+++ src/services/random/hooks.ts
@@ -1,0 +1,7 @@
+import { randomUUID } from "crypto";
+import { useState } from "react";
+
+export const useId = (): string => {
+	const [id] = useState(randomUUID());
+	return id;
+};
