+++ src/services/string/conversion.ts
@@ -12,3 +12,20 @@ export const pxToNum = (pixels: string) => {
 export const numToPx = (num: number) => {
 	return `${num}px`;
 };
+
+export const dateTimeToString = (dateTime: number) => {
+	const date = new Date(dateTime);
+	const dateString = date.toLocaleString("en-US", {
+		month: "short",
+		day: "numeric",
+		year: "numeric",
+		hour: "numeric",
+		minute: "numeric",
+		hour12: true,
+	});
+	const commaIndex = dateString.lastIndexOf(",");
+	return (
+		dateString.substring(0, commaIndex) +
+		dateString.substring(commaIndex + 1)
+	);
+};
